// server/src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const OrderStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  SHIPPED: 'SHIPPED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]